+++ NEWS.md
@@ -1,5 +1,13 @@
 # prediction 0.2.9
 
+* Added `prediction.bruto()`method for "bruto" objects from **mda**. (#1)
+* Added `prediction.fda()`method for "fda" objects from **mda**. (#1)
+* Added `prediction.mars()`method for "mars" objects from **mda**. (#1)
+* Added `prediction.mda()`method for "mda" objects from **mda**. (#1)
+* Added `prediction.polyreg()`method for "polyreg" objects from **mda**. (#1)
+
+# prediction 0.2.9
+
 * Added `prediction.speedglm()` and `prediction.speedlm()` methods for "speedglm" and "speedlm" objects from **speedglm**. (#1)
 * Added `prediction.bigLm()` method for "bigLm" objects from **bigFastlm**. (#1)
 * Added `prediction.biglm()` and `prediction.bigglm()` methods for "biglm" and "bigglm" objects from **biglm**, including those based by `"ffdf"` from **ff**. (#1)